Add redirect print to capa main
