+++ capa/main.py
@@ -53,6 +53,7 @@ from capa.helpers import (
     get_file_taste,
     get_auto_format,
     log_unsupported_os_error,
+    redirecting_print_to_tqdm,
     log_unsupported_arch_error,
     log_unsupported_format_error,
 )
@@ -251,38 +252,39 @@ def find_capabilities(ruleset: RuleSet, extractor: FeatureExtractor, disable_pro
         "library_functions": {},
     }  # type: Dict[str, Any]
 
-    pbar = tqdm.tqdm
-    if disable_progress:
-        # do not use tqdm to avoid unnecessary side effects when caller intends
-        # to disable progress completely
-        def pbar(s, *args, **kwargs):
-            return s
+    with redirecting_print_to_tqdm(disable_progress):
+        pbar = tqdm.tqdm
+        if disable_progress:
+            # do not use tqdm to avoid unnecessary side effects when caller intends
+            # to disable progress completely
+            def pbar(s, *args, **kwargs):
+                return s
 
-    functions = list(extractor.get_functions())
-    n_funcs = len(functions)
+        functions = list(extractor.get_functions())
+        n_funcs = len(functions)
 
-    pb = pbar(functions, desc="matching", unit=" functions", postfix="skipped 0 library functions")
-    for f in pb:
-        if extractor.is_library_function(f.address):
-            function_name = extractor.get_function_name(f.address)
-            logger.debug("skipping library function 0x%x (%s)", f.address, function_name)
-            meta["library_functions"][f.address] = function_name
-            n_libs = len(meta["library_functions"])
-            percentage = round(100 * (n_libs / n_funcs))
-            if isinstance(pb, tqdm.tqdm):
-                pb.set_postfix_str(f"skipped {n_libs} library functions ({percentage}%)")
-            continue
+        pb = pbar(functions, desc="matching", unit=" functions", postfix="skipped 0 library functions")
+        for f in pb:
+            if extractor.is_library_function(f.address):
+                function_name = extractor.get_function_name(f.address)
+                logger.debug("skipping library function 0x%x (%s)", f.address, function_name)
+                meta["library_functions"][f.address] = function_name
+                n_libs = len(meta["library_functions"])
+                percentage = round(100 * (n_libs / n_funcs))
+                if isinstance(pb, tqdm.tqdm):
+                    pb.set_postfix_str(f"skipped {n_libs} library functions ({percentage}%)")
+                continue
 
-        function_matches, bb_matches, insn_matches, feature_count = find_code_capabilities(ruleset, extractor, f)
-        meta["feature_counts"]["functions"][f.address] = feature_count
-        logger.debug("analyzed function 0x%x and extracted %d features", f.address, feature_count)
+            function_matches, bb_matches, insn_matches, feature_count = find_code_capabilities(ruleset, extractor, f)
+            meta["feature_counts"]["functions"][f.address] = feature_count
+            logger.debug("analyzed function 0x%x and extracted %d features", f.address, feature_count)
 
-        for rule_name, res in function_matches.items():
-            all_function_matches[rule_name].extend(res)
-        for rule_name, res in bb_matches.items():
-            all_bb_matches[rule_name].extend(res)
-        for rule_name, res in insn_matches.items():
-            all_insn_matches[rule_name].extend(res)
+            for rule_name, res in function_matches.items():
+                all_function_matches[rule_name].extend(res)
+            for rule_name, res in bb_matches.items():
+                all_bb_matches[rule_name].extend(res)
+            for rule_name, res in insn_matches.items():
+                all_insn_matches[rule_name].extend(res)
 
     # collection of features that captures the rule matches within function, BB, and instruction scopes.
     # mapping from feature (matched rule) to set of addresses at which it matched.
